+++ src/components/super/i-static-page/modules/theme/index.ts
@@ -11,6 +11,8 @@
  * @packageDocumentation
  */
 
-export { default } from 'components/super/i-static-page/modules/theme/factory';
-export { default as ThemeManager } from 'components/super/i-static-page/modules/theme/theme-manager';
 export * from 'components/super/i-static-page/modules/theme/const';
+export { default } from 'components/super/i-static-page/modules/theme/factory';
+
+export { default as ThemeManager } from 'components/super/i-static-page/modules/theme/theme-manager';
+export { default as SystemThemeExtractorWeb } from 'components/super/i-static-page/modules/theme/system-theme-extractor/engines/web/engine';