+++ src/core/data/interface/types.ts
@@ -13,6 +13,7 @@ import type { Provider } from 'core/data/interface';
 export * from '@v4fire/core/core/data/interface/types';
 
 export interface ProviderOptions extends SuperProviderOptions {
+	id: string;
 	remoteState?: State;
 	i18n?: i18nFactory;
 }